refactor(cli): split commands/mod.rs god-module into per-domain siblings
The 1.7k-line commands module index held ~45 command handlers plus print
and format helpers. Move them into cohesive sibling files (search, clones,
index_ops, models, memory, hooks, config_info, runtime_env, format) and
reduce mod.rs to a curated index that re-exports the command surface main.rs
dispatches to.

No behavior change: pure move + call-site migration, no re-export shims.
