+++ crates/rag-rat-cli/src/commands/runtime_env.rs
@@ -1,0 +1,28 @@
+//! Embedding-runtime environment setup, split out of the `commands` god-module. Applied once at
+//! process startup (from `main` and the `init` wizard) before rag-rat builds its Tokio runtime or
+//! initializes FastEmbed/ONNX.
+use std::env;
+
+use rag_rat_core::config::EmbeddingRuntimeConfig;
+
+pub(crate) fn apply_embedding_runtime_env(runtime: &EmbeddingRuntimeConfig) {
+    // `ort_threads` is applied via fastembed's session `with_intra_threads` (see
+    // FastEmbedEmbedder::new), not an env var — ONNX Runtime does not read `ORT_NUM_THREADS`.
+    // `omp_threads` IS effective: Microsoft's prebuilt ORT is OpenMP-based and honors
+    // `OMP_NUM_THREADS`, so it is the real thread lever for the default binaries.
+    set_env_if_absent("OMP_NUM_THREADS", runtime.omp_threads);
+}
+
+pub(crate) fn set_env_if_absent(key: &str, value: Option<u32>) {
+    let Some(value) = value else {
+        return;
+    };
+    if env::var_os(key).is_some() {
+        return;
+    }
+    // This is called at process startup before rag-rat creates its Tokio runtime or initializes
+    // FastEmbed/ONNX. CLI-provided environment variables intentionally take precedence.
+    unsafe {
+        env::set_var(key, value.to_string());
+    }
+}
